fixes, adjusted GUI
